move Maps into a folder
