+++ frontend/src/components/Maps/MapCard.js
